fixing problems and adding fonctions
ajouter , supprimer emp , consulter emp, editer emp , plus paggination sur consuler liste employee

+++ hr-management-frontend-angular/src/app/consulter/consulter.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { DataService } from '../data.service';
+import { Router } from '@angular/router';
 
 
 @Component({
@@ -10,28 +11,63 @@ import { DataService } from '../data.service';
 export class ConsulterComponent implements OnInit {
 
 
-  constructor(private empService: DataService) { }
+  constructor(private empService: DataService, private router:Router) { }
 
-  employees ;
-
+  public employees:any ;
+  public size:number=6;
+  public currentPage:number=0;
+  public totalPages:number;
+  public pages:Array<number>;
   ngOnInit() {
+ // l'appel du fonction getAllPersonnels ( consulter les employees)
 
-    this.empService.getAllPersonnels()
+    this.empService.getAllPersonnels(this.currentPage,this.size)
       .subscribe(data => {
-      this.employees = data;
+      
+      this.totalPages=data["page"].totalPages;
+      this.pages=new Array<number>(this.totalPages);
+     this.employees = data;
+
       }, err => {
         console.log(err)
       })
 
   }
-
-  onDeletePers(emp) {
-
-    this.empService.deleteP(emp.empNo).subscribe(data => {
+ // l'appel du fonction deleteP => pour le button supprimer un employe
+  onDeleteEmp(emp) {
+    let conf= confirm("Etes vous sure ?");
+    if (conf) {
+      this.empService.deleteP(emp.empNo).subscribe(data => {
     this.employees = data;
+    this.ngOnInit();
     }, err => {
       console.log(err)
     })
+    }
+    
   }
+
+// recupperer l'id pour editer un employe 
+onEditEmp1(emp){
+  let url =emp._links.self.href;
+this.router.navigateByUrl("/editer/"+btoa(url) );
+}
+
+//Consulter un employe 
+onConsultEmp1(emp){
+
+  let url =emp._links.self.href;
+this.router.navigateByUrl("/consulterEmp/"+btoa(url) );
+
+}
+
+// la fonction qui retourne la page consulter apres avoir 
+// cliquer sur la pagination :
+  onPageEmp(i){
+    this.currentPage=i;
+    this.ngOnInit();
+  }
+
+ 
 
 }
